feat: implement navigation to movie details screen

+++ feature/tmdb/impl/src/main/java/com/majorkik/tmdb/impl/repository/MovieDetailsRepositoryImpl.kt
@@ -8,7 +8,7 @@ import com.majorkik.tmdb.impl.network.safeRequest
 import com.majorkik.tmdb.impl.respone.toDomainModel
 
 internal class MovieDetailsRepositoryImpl(private val api: ApiService) : MovieDetailsRepository {
-    override suspend fun fetchMovieDetails(id: Long): NetworkResult<MovieDetails, String> {
+    override suspend fun fetchMovieDetails(id: Int): NetworkResult<MovieDetails, String> {
         return safeRequest(
             call = { api.getMovieById(id = id, appendToResponse = "images") },
             onSuccess = { it?.toDomainModel() },
